+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/add_ms_statistics_original.sql
@@ -1,0 +1,26 @@
+CREATE TABLE [dbo].[t6](
+    [col1] [int] NULL,
+    [col2] [int] NULL,
+    [col3] [int] NULL
+) ON [PRIMARY]
+GO
+
+CREATE TABLE [dbo].[t4](
+    [col1] [int] NULL,
+    [col2] [int] NULL,
+    [col3] [int] NULL
+) ON [PRIMARY]
+GO
+
+CREATE PARTITION FUNCTION myRangePF1 (datetime2(0))
+    AS RANGE RIGHT FOR VALUES ('2022-04-01', '2022-05-01', '2022-06-01') ;
+GO
+
+CREATE PARTITION SCHEME myRangePS1
+    AS PARTITION myRangePF1
+    ALL TO ('PRIMARY') ;
+GO
+
+CREATE TABLE dbo.PartitionTable (col1 datetime2(0) PRIMARY KEY, col2 char(10))
+    ON myRangePS1 (col1) ;
+GO