+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/change_ms_table_with_stat_usr_diff.sql
@@ -1,0 +1,17 @@
+-- DEPCY: This STATISTICS depends on the COLUMN: [dbo].[t1].[col3]
+
+DROP STATISTICS [dbo].[t1].[stat1]
+GO
+
+-- DEPCY: This STATISTICS depends on the COLUMN: [dbo].[t1].[col3]
+
+DROP STATISTICS [dbo].[t1].[stat2]
+GO
+
+ALTER TABLE [dbo].[t1]
+	DROP COLUMN [col3]
+GO
+
+CREATE STATISTICS stat1 ON [dbo].[t1] ([col1], [col2])
+WHERE ([col2]=(3))
+GO