+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ms_statistics_diff.sql
@@ -1,0 +1,8 @@
+DROP STATISTICS [dbo].[t6].[stat1]
+GO
+
+DROP STATISTICS [dbo].[t4].[stat1]
+GO
+
+DROP TABLE [dbo].[t4]
+GO