+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ms_statistics_new.sql
@@ -1,0 +1,6 @@
+CREATE TABLE [dbo].[t6](
+    [col1] [int] NULL,
+    [col2] [int] NULL,
+    [col3] [int] NULL
+) ON [PRIMARY]
+GO